+++ system/syslog_windows.go
@@ -1,0 +1,10 @@
+// +build windows
+
+package system
+
+import "errors"
+
+// EnableSyslog is not supported on windows
+func EnableSyslog() error {
+	return errors.New("Syslog not supported on windows")
+}